Ensure that target is properly unloaded on free.
Otherwise if zypp is then again initialized by another thread,
this thread will do internally in libzypp unload of old target and
result is segafault or memory misalignment.
It is not bullet proof, but I do not find yet way how to say that even
after freeing target, it should be again created only in identical
thread. Only we can do by holding static zypp pointer and do not allow
to free it.